If a file is big enough, and empty enough, a free range can overflow 32-bits; change free-range specifications from off_t/uint32_t pairs to off_t/off_t pairs.
According the Single UNIX Specification, Version 2, off_t is a signed
type, use intmax_t/PRIdMAX instead of uintmax_t/PRIuMAX to print it.

+++ src/include/block.h
@@ -39,7 +39,7 @@
  */
 struct __wt_free {
 	off_t	 off;				/* File offset */
-	uint32_t size;				/* Size */
+	off_t	 size;				/* Size */
 
 	uint8_t	 depth;				/* Skip list depth */
 
@@ -56,7 +56,7 @@ struct __wt_free {
  *	Encapsulation of a block size skiplist entry.
  */
 struct __wt_size {
-	uint32_t size;				/* Size */
+	off_t	 size;				/* Size */
 
 	uint8_t	 depth;				/* Skip list depth */
